# src/proyecto/proyectoFinCurso/Services/ApiClients/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/users/{id}":{"get":{"tags":["user-controller"],"summary":"Get user by ID","description":"Retrieve a user record by its unique identifier.","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["user-controller"],"summary":"Update user by ID","description":"Update the user record identified by its ID with new data.","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["user-controller"],"summary":"Delete user by ID","description":"Remove the user record identified by its ID.","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/table/{id}":{"get":{"tags":["table-controller"],"summary":"Get table by ID","description":"Retrieve a table record by its unique identifier.","operationId":"getTableById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TableResponse"}}}}}},"put":{"tags":["table-controller"],"summary":"Update table by ID","description":"Update the table record identified by its ID with new data.","operationId":"updateTable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TableResponse"}}}}}},"delete":{"tags":["table-controller"],"summary":"Delete table by ID","description":"Remove the table record identified by its ID.","operationId":"deleteTable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/product/{id}":{"get":{"tags":["product-controller"],"summary":"Get product by ID","description":"Retrieve a product record by its unique identifier.","operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"put":{"tags":["product-controller"],"summary":"Update product by ID","description":"Update the product record identified by its ID with new data.","operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}},"delete":{"tags":["product-controller"],"summary":"Delete product by ID","description":"Remove the product record identified by its ID.","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/order/{id}":{"get":{"tags":["order-controller"],"summary":"Get order by ID","description":"Retrieve an order by its unique identifier.","operationId":"getOrderById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}},"put":{"tags":["order-controller"],"summary":"Update order by ID","description":"Update the order identified by its ID with new data.","operationId":"updateOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}},"delete":{"tags":["order-controller"],"summary":"Delete order by ID","description":"Remove the order identified by its ID.","operationId":"deleteOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/order-state-history/{id}":{"get":{"tags":["order-state-history-controller"],"summary":"Get order state history by ID","description":"Retrieve an order state history record by its unique identifier.","operationId":"getOrderStateHistoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderStateHistoryResponse"}}}}}},"put":{"tags":["order-state-history-controller"],"summary":"Update order state history by ID","description":"Update the order state history record identified by its ID with new data.","operationId":"updateOrderStateHistory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStateHistoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderStateHistoryResponse"}}}}}},"delete":{"tags":["order-state-history-controller"],"summary":"Delete order state history by ID","description":"Remove the order state history record identified by its ID.","operationId":"deleteOrderStateHistory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/order-detail/{id}":{"get":{"tags":["order-detail-controller"],"summary":"Get order detail by ID","description":"Retrieve an order detail record by its unique identifier.","operationId":"getOrderDetailById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDetailResponse"}}}}}},"put":{"tags":["order-detail-controller"],"summary":"Update order detail by ID","description":"Update the order detail identified by its ID with new data.","operationId":"updateOrderDetail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDetailRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDetailResponse"}}}}}},"delete":{"tags":["order-detail-controller"],"summary":"Delete order detail by ID","description":"Remove the order detail record identified by its ID.","operationId":"deleteOrderDetail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/inventory/{id}":{"get":{"tags":["inventory-controller"],"summary":"Get inventory item by ID","description":"Retrieve an inventory record by its unique identifier.","operationId":"getInventoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventoryResponse"}}}}}},"put":{"tags":["inventory-controller"],"summary":"Update inventory item by ID","description":"Update the inventory record identified by its ID with new data.","operationId":"updateInventory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventoryResponse"}}}}}},"delete":{"tags":["inventory-controller"],"summary":"Delete inventory item by ID","description":"Remove the inventory record identified by its ID.","operationId":"deleteInventory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/users":{"post":{"tags":["user-controller"],"summary":"Create a new user","description":"Create a new user with the provided details.","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/table":{"post":{"tags":["table-controller"],"summary":"Create a new table","description":"Create a new table entry with the provided details.","operationId":"createTable","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TableResponse"}}}}}}},"/product":{"post":{"tags":["product-controller"],"summary":"Create a new product","description":"Create a new product entry with the provided details.","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}}}}},"/order":{"post":{"tags":["order-controller"],"summary":"Create a new order","description":"Create a new order with the provided details.","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}}}}},"/order-state-history":{"post":{"tags":["order-state-history-controller"],"summary":"Create a new order state history record","description":"Create a new order state history entry with the provided details.","operationId":"createOrderStateHistory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStateHistoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderStateHistoryResponse"}}}}}}},"/order-detail":{"post":{"tags":["order-detail-controller"],"summary":"Create a new order detail","description":"Create a new order detail entry with the provided details.","operationId":"createOrderDetail","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDetailRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDetailResponse"}}}}}}},"/inventory":{"post":{"tags":["inventory-controller"],"summary":"Create a new inventory record","description":"Create a new inventory entry with the provided details.","operationId":"createInventory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InventoryResponse"}}}}}}}},"components":{"schemas":{"UserRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":45,"minLength":0},"surnames":{"type":"string","maxLength":45,"minLength":0},"username":{"type":"string","maxLength":45,"minLength":0},"password":{"type":"string","maxLength":45,"minLength":0},"email":{"type":"string","maxLength":45,"minLength":0},"phone":{"type":"string","maxLength":45,"minLength":0}},"required":["name","password","phone","surnames","username"]},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":45,"minLength":0},"surnames":{"type":"string","maxLength":45,"minLength":0},"username":{"type":"string","maxLength":45,"minLength":0},"email":{"type":"string","maxLength":45,"minLength":0},"phone":{"type":"string","maxLength":9,"minLength":0}},"required":["name","phone","surnames","username"]},"TableRequest":{"type":"object","properties":{"num_Table":{"type":"integer","format":"int32"},"num_Customers":{"type":"integer","format":"int32"},"waiter":{"$ref":"#/components/schemas/UserRequest"},"state":{"type":"string","maxLength":45,"minLength":0}},"required":["num_Customers","num_Table","state","waiter"]},"TableResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"num_Table":{"type":"integer","format":"int32"},"num_Customers":{"type":"integer","format":"int32"},"waiter":{"$ref":"#/components/schemas/UserResponse"},"state":{"type":"string","maxLength":45,"minLength":0}},"required":["num_Customers","num_Table","state","waiter"]},"ProductRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":45,"minLength":0},"price":{"type":"integer","format":"int32"},"category":{"type":"string","maxLength":45,"minLength":0},"desc":{"type":"string"}},"required":["name","price"]},"ProductResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":45,"minLength":0},"price":{"type":"integer","format":"int32"},"category":{"type":"string","maxLength":45,"minLength":0},"desc":{"type":"string"}},"required":["name","price"]},"OrderRequest":{"type":"object","properties":{"table":{"$ref":"#/components/schemas/TableRequest"},"date":{"type":"string","format":"date-time"},"state":{"type":"string","maxLength":45,"minLength":0}},"required":["date","state","table"]},"OrderResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"table":{"$ref":"#/components/schemas/TableResponse"},"date":{"type":"string","format":"date-time"},"state":{"type":"string","maxLength":45,"minLength":0}},"required":["date","state","table"]},"OrderStateHistoryRequest":{"type":"object","properties":{"order":{"$ref":"#/components/schemas/OrderRequest"},"oldState":{"type":"string","maxLength":45,"minLength":0},"newState":{"type":"string","maxLength":45,"minLength":0},"date":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/UserRequest"}},"required":["date","newState","order","user"]},"OrderStateHistoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"order":{"$ref":"#/components/schemas/OrderResponse"},"oldState":{"type":"string","maxLength":45,"minLength":0},"newState":{"type":"string","maxLength":45,"minLength":0},"date":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/UserResponse"}},"required":["date","newState","order","user"]},"OrderDetailRequest":{"type":"object","properties":{"order":{"$ref":"#/components/schemas/OrderRequest"},"product":{"$ref":"#/components/schemas/ProductRequest"},"amount":{"type":"integer","format":"int32"}},"required":["amount","order","product"]},"OrderDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"order":{"$ref":"#/components/schemas/OrderResponse"},"product":{"$ref":"#/components/schemas/ProductResponse"},"amount":{"type":"integer","format":"int32"}},"required":["amount","order","product"]},"InventoryRequest":{"type":"object","properties":{"product":{"$ref":"#/components/schemas/ProductRequest"},"amount":{"type":"integer","format":"int32"}},"required":["product"]},"InventoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"product":{"$ref":"#/components/schemas/ProductResponse"},"amount":{"type":"integer","format":"int32"}},"required":["product"]}}}}